+++ boxy-core/src/main/java/boxy/core/model/ConsumerGroup.java
@@ -14,6 +14,6 @@ public record ConsumerGroup(
         int releaseDeadline,
         int activePartitions,
         int activeWorkers,
-        int activeWorkersWeight,
+        double activeWorkersWeight,
         Instant lastUpdated) {
 }
