+++ boxy-core/src/main/java/boxy/core/model/Worker.java
@@ -6,7 +6,7 @@ public record Worker(
         long id,
         String nodeId,
         long consumerGroupId,
-        int weight,
+        double weight,
         Instant lastHeartbeat,
         double heartbeatInterval,
         Instant heartbeatDeadline) {
